[DT-21] implement install manager for DT

+++ install-manager/src/main/java/cm/aptoide/pt/install_manager/App.kt
@@ -17,7 +17,7 @@ import cm.aptoide.pt.install_manager.dto.InstallPackageInfo
 interface App<D> {
   val packageName: String
   val installedVersionName: String?
-  val installedVersionCode: Int?
+  val installedVersionCode: Long?
   val details: D?
 
   /**
